Move LP setup to correct file
