--- src/scheduling/extended_schedule.py
@@ -1,239 +1,0 @@
-import pulp
-
-from src.common import CircuitJob, ScheduledJob
-from src.provider import Accelerator
-
-from .calculate_makespan import calculate_makespan
-from .extract_schedule import extract_info_schedule, extract_executable_schedule
-from .setup_lp import set_up_base_lp
-from .solve_lp import solve_lp
-from .types import JobResultInfo, LPInstance, PTimes, STimes
-
-
-def generate_extended_lp(
-    lp_instance: LPInstance,
-    process_times: PTimes,
-    setup_times: STimes,
-    big_m: int = 1000,
-) -> LPInstance:
-    p_times = pulp.makeDict(
-        [lp_instance.jobs[1:], lp_instance.machines],
-        process_times,
-        0,
-    )
-    s_times = pulp.makeDict(
-        [lp_instance.jobs, lp_instance.jobs, lp_instance.machines],
-        setup_times,
-        0,
-    )
-    # decision variables
-    y_ijk = pulp.LpVariable.dicts(
-        "y_ijk",
-        (lp_instance.jobs, lp_instance.jobs, lp_instance.machines),
-        cat="Binary",
-    )
-    a_ij = pulp.LpVariable.dicts(
-        "a_ij", (lp_instance.jobs, lp_instance.jobs), cat="Binary"
-    )  # a: Job i ends before job j starts
-    b_ij = pulp.LpVariable.dicts(
-        "b_ij", (lp_instance.jobs, lp_instance.jobs), cat="Binary"
-    )  # b: Job i ends before job j ends
-    d_ijk = pulp.LpVariable.dicts(
-        "d_ijk",
-        (lp_instance.jobs, lp_instance.jobs, lp_instance.machines),
-        cat="Binary",
-    )  # d: Job i and  j run on the same machine
-    e_ijlk = pulp.LpVariable.dicts(
-        "e_ijlk",
-        (lp_instance.jobs, lp_instance.jobs, lp_instance.jobs, lp_instance.machines),
-        cat="Binary",
-    )
-
-    for job in lp_instance.jobs[1:]:
-        lp_instance.problem += (  # (4)
-            pulp.lpSum(
-                y_ijk[job_j][job][machine]
-                for machine in lp_instance.machines
-                for job_j in lp_instance.jobs
-            )
-            >= 1  # each job has a predecessor
-        )
-        lp_instance.problem += lp_instance.c_j[job] >= lp_instance.s_j[  # (7)
-            job
-        ] + pulp.lpSum(
-            lp_instance.x_ik[job][machine] * p_times[job][machine]
-            for machine in lp_instance.machines
-        ) + pulp.lpSum(
-            y_ijk[job_j][job][machine] * s_times[job_j][job][machine]
-            for machine in lp_instance.machines
-            for job_j in lp_instance.jobs
-        )
-        for machine in lp_instance.machines:
-            lp_instance.problem += (  # predecessor (6)
-                lp_instance.x_ik[job][machine]
-                >= pulp.lpSum(y_ijk[job_j][job][machine] for job_j in lp_instance.jobs)
-                / big_m
-            )
-            lp_instance.problem += (  # successor
-                lp_instance.x_ik[job][machine]
-                >= pulp.lpSum(y_ijk[job][job_j][machine] for job_j in lp_instance.jobs)
-                / big_m
-            )
-            lp_instance.problem += (  # (5)
-                lp_instance.z_ikt[job][machine][0] == y_ijk["0"][job][machine]
-            )
-        for job_j in lp_instance.jobs:
-            lp_instance.problem += (
-                lp_instance.c_j[job_j]
-                + (
-                    pulp.lpSum(
-                        y_ijk[job_j][job][machine] for machine in lp_instance.machines
-                    )
-                    - 1
-                )
-                * big_m
-                <= lp_instance.s_j[job]
-            )
-
-    # Extended constraints
-    for job in lp_instance.jobs[1:]:
-        for job_j in lp_instance.jobs[1:]:
-            if job == job_j:
-                lp_instance.problem += a_ij[job][job_j] == 0
-                lp_instance.problem += b_ij[job][job_j] == 0
-                continue
-            lp_instance.problem += (
-                a_ij[job][job_j]
-                >= (lp_instance.s_j[job_j] - lp_instance.c_j[job]) / big_m
-            )
-            lp_instance.problem += (
-                b_ij[job][job_j]
-                >= (lp_instance.c_j[job_j] - lp_instance.c_j[job]) / big_m
-            )
-            for machine in lp_instance.machines:
-                lp_instance.problem += (
-                    d_ijk[job][job_j][machine]
-                    >= lp_instance.x_ik[job][machine]
-                    + lp_instance.x_ik[job_j][machine]
-                    - 1
-                )
-                for job_l in lp_instance.jobs[1:]:
-                    lp_instance.problem += (
-                        e_ijlk[job][job_j][job_l][machine]
-                        >= b_ij[job][job_l]
-                        + a_ij[job_l][job_j]
-                        + d_ijk[job][job_j][machine]
-                        + d_ijk[job][job_l][machine]
-                        - 3
-                    )
-
-    for job in lp_instance.jobs[1:]:
-        for job_j in lp_instance.jobs[1:]:
-            for machine in lp_instance.machines:
-                lp_instance.problem += (
-                    y_ijk[job][job_j][machine]
-                    >= a_ij[job][job_j]
-                    + (
-                        pulp.lpSum(
-                            e_ijlk[job][job_j][job_l][machine]
-                            for job_l in lp_instance.jobs[1:]
-                        )
-                        / big_m
-                    )
-                    + d_ijk[job][job_j][machine]
-                    - 2
-                )
-    return lp_instance
-
-
-def generate_extended_schedule(
-    lp_instance: LPInstance,
-    process_times: PTimes,
-    setup_times: STimes,
-    big_m: int = 1000,
-) -> tuple[float, list[JobResultInfo]]:
-    """Generates the extended schedule for the given jobs and accelerators using a complex MILP.
-
-    First generates the schedule using MILP  and then calculates the makespan
-    by executing the schedule with the correct p_ij and s_ij values.
-
-    Args:
-        lp_instance (LPInstance): The base LP instance.
-        process_times (PTimes): The process times for each job on each machine.
-        setup_times (STimes): The setup times for each job on each machine.
-        big_m (int, optional): Metavariable for the LP. Defaults to 1000.
-
-    Returns:
-        tuple[float, list[JobResultInfo]]: List of jobs with their assigned machine and
-            start and completion times.
-    """
-    lp_instance = generate_extended_lp(lp_instance, process_times, setup_times, big_m)
-    _, jobs = extract_info_schedule(solve_lp(lp_instance))
-    return calculate_makespan(lp.instance, jobs, process_times, setup_times), jobs
-
-
-def generate_extended_schedule_provider(
-    jobs: list[CircuitJob],
-    accelerators: list[Accelerator],
-    big_m: int = 1000,
-    t_max: int = 2**7,
-) -> list[ScheduledJob]:
-    """Calclulate a schedule for the given jobs and accelerators based on the extended MILP.
-
-    The extended MILP includes the sequence dependent setup time between jobs.
-    This depends on the unique pre- and successor condition described in the paper.
-    Args:
-        jobs (list[CircuitJob]): The list of jobs to run.
-        accelerators (list[Accelerator]): The list of available accelerators.
-        big_m (int, optional): M hepler for LP. Defaults to 1000.
-        t_max (int, optional): Max number of Timesteps. Defaults to 2**7.
-    Returns:
-        list[ScheduledJob]: A list of Jobs scheduled to accelerators.
-    """
-    lp_instance = set_up_base_lp(jobs, accelerators, big_m, list(range(t_max)))
-
-    # additional parameters
-    p_times = pulp.makeDict(
-        [lp_instance.jobs, lp_instance.machines],
-        _get_processing_times(jobs, accelerators),
-        0,
-    )
-    # TODO check if this works correctly for job "0"
-    s_times = pulp.makeDict(
-        [lp_instance.jobs, lp_instance.jobs, lp_instance.machines],
-        _get_setup_times(jobs, accelerators),
-        0,
-    )
-    lp_instance = generate_extended_lp(
-        lp_instance=lp_instance, process_times=p_times, setup_times=s_times
-    )
-
-    return extract_executable_schedule(solve_lp(lp_instance), jobs, accelerators)
-
-
-def _get_setup_times(
-    base_jobs: list[CircuitJob], accelerators: list[Accelerator]
-) -> list[list[list[float]]]:
-    return [
-        [
-            [
-                qpu.compute_setup_time(job_i.circuit, job_j.circuit)
-                for qpu in accelerators
-            ]
-            for job_i in base_jobs
-            if job_i.circuit is not None
-        ]
-        for job_j in base_jobs
-        if job_j.circuit is not None
-    ]
-
-
-def _get_processing_times(
-    base_jobs: list[CircuitJob],
-    accelerators: list[Accelerator],
-) -> list[list[float]]:
-    return [
-        [qpu.compute_processing_time(job.circuit) for qpu in accelerators]
-        for job in base_jobs
-        if job.circuit is not None
-    ]
